Restructure of some included profiles.
First thing this does is add some filename conventions. We already have wiimote_ for stuff that only applies to the wiimotes. We have ps4_ that applies to the dualshock 4s. I propose we use mg_ for things that affect the virtual gamepads. dpad_to_ls_on is now mg_dpad_to_ls.

In addition, I have gotten rid of dpad_to_ls_off and replaced it with a profile called mg_reset. We only have the dpad-to-ls tweak right now, but we could conceivably add more, and stack them on top of each other. This single profile load would then revert all those changes to normal. I have tested it on my end--though not extensively--and it currently does what I expect it to.